Hydrate WVUI search
Pass the initial state of the server-side rendered search to Vue.js
search. This requires an exact DOM structure that _would_ be generated
by the Node.js Vue.js server renderer for the Vue.js client renderer to
work without throwing an error.

The input tag was chosen for hydration as opposed to the entirety of the
search form to minimize the risk that a script would modify the DOM
structure which could cause a runtime error. The PHP implementation
generates exactly an input element so it is unexpected for this to
shift. See buildSearchProps() in core/includes/skins/SkinMustache.php.
Any listeners added to that input will be preserved by Vue.js as proper
client hydration maintains the existing DOM.

See additional complications and concerns identified in T260065,
https://github.com/wikimedia/wvui/pull/73, and
https://github.com/wikimedia/wvui/pull/67.

Also, add `build:watch` NPM script as overriding `-p` doesn't seem to
work with vue-loader.

Bug: T260065
Bug: T257647
Change-Id: Ic2f8661bd7d3d98f4be8e7d545f04236d1ed1aee