Set Vector search variables relative to @font-size-base factor and increase max-width
T270202 shows the correct dimensions of the search component having a
max-width of 500px and min-width of 350px but since we set the font size
of `#p-search` to `@font-size-base` [1], I suspect the search related
variables need to account for this factor as well.

This increases the max-width of search to 500px per the spec.

[1] https://github.com/wikimedia/Vector/blob/30eb683a709f8b4dc06c64571ce3c39a3c885e01/resources/skins.vector.styles/VueEnhancedSearchBox.less#L23

Bug: T270202
Change-Id: I13ad550734e8a5347ed70e4b3c33102b4a13bde7